requirements.txt
pyside6
xlsxwriter
pywin32; sys_platform == 'win32'
python-docx